requirements.txt
alpaca-py
python-dotenv
pandas
numpy
yfinance
polygon-api-client